Changed ui of homepage and recipe links
I moved the homepage link from the bottom of the recipe pages to right
under the image for easier access in case the user want to quickly go
back in case they misclicked on a link.

I changed the format of the recipe links from a single line to an
unordered list for to make the page more appealing.